Added Gradle support, making the project less IDE-dependant
