--- Matrica/.settings/org.eclipse.jdt.ui.prefs
@@ -1,3 +1,0 @@
-eclipse.preferences.version=1
-formatter_profile=_mine
-formatter_settings_version=12
