--- Matrica/src/ChooseListener.java
@@ -1,4 +1,0 @@
-
-public interface ChooseListener{
-	public void chooseOption(int choose);
-}
