--- Matrica/src/MatricaLayout.java
@@ -1,10 +1,0 @@
-import javax.swing.JPanel;
-
-
-public class MatricaLayout extends JPanel{
-	
-	public MatricaLayout(){
-		// Perdor card layout *
-	}
-	
-}
