--- Matrica/src/launch.java
@@ -1,14 +1,0 @@
-import javax.swing.SwingUtilities;
-
-
-public class launch {
-	
-	public static void main(String[] args){
-		SwingUtilities.invokeLater(new Runnable() {
-			public void run() {
-				new MainFrame();
-			}
-		});
-	}
-	
-}
