+++ README.md
@@ -1,2 +1,6 @@
 # Matrica
-A sample application for solving matrix equations
+A sample application for solving matrix equations. You can run the application by executing:
+
+```bash
+# ./gradlew run
+```
