+++ src/main/java/matrix/ChooseListener.java
@@ -1,0 +1,5 @@
+package matrix;
+
+public interface ChooseListener {
+    void chooseOption(int choose);
+}
