+++ src/main/java/matrix/MatricaLayout.java
@@ -1,0 +1,11 @@
+package matrix;
+
+import javax.swing.JPanel;
+
+public class MatricaLayout extends JPanel {
+
+    public MatricaLayout() {
+        // Perdor card layout *
+    }
+
+}
