+++ src/main/java/matrix/launch.java
@@ -1,0 +1,15 @@
+package matrix;
+
+import javax.swing.SwingUtilities;
+
+public class launch {
+
+    public static void main(String[] args) {
+        SwingUtilities.invokeLater(new Runnable() {
+            public void run() {
+                new MainFrame();
+            }
+        });
+    }
+
+}
